pyproject.toml
[tool.poetry]
name = "covid_info"
version = "0.1.0"
description = ""
authors = ["XuQiang <xq_work@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
rich = "^12.6.0"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
